build: split style_test into a proper lint package
build/style_test.go is rather substantial and deserves its own package.
Move it to pkg/lint. This has the nice side effect of ensuring the
linter lints itself.
